Refactor : 프로젝트 구조를 계층형 구조로 수정
