--- src/main/java/com/tumbloom/tumblerin/app/User/RoleType.java
@@ -1,5 +1,0 @@
-package com.tumbloom.tumblerin.app.User;
-
-public enum RoleType {
-    USER,ADMIN
-}
