+++ src/main/java/com/tumbloom/tumblerin/app/domain/RoleType.java
@@ -1,0 +1,5 @@
+package com.tumbloom.tumblerin.app.domain;
+
+public enum RoleType {
+    USER,ADMIN
+}
