+++ src/main/java/com/tumbloom/tumblerin/app/dto/Authdto/LoginResponseDTO.java
@@ -1,8 +1,7 @@
-package com.tumbloom.tumblerin.app.User.dto;
+package com.tumbloom.tumblerin.app.dto.Authdto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.Setter;
 
 @Getter
 @AllArgsConstructor
